cc: fix char literal signedness and static local struct placement
Bug AA - Character literal signedness: '\x80' produced 128 (unsigned)
instead of -128 (signed). pcc converted Rust char to i64 without
sign-extension from 8 bits. Fixed by using `*c as u8 as i8 as i64` in
all 4 CharLit-to-integer paths. This broke CPython's _pickle module
where enum opcodes >= 0x80 couldn't match switch cases.

Bug AB - Static local struct on stack: linearize_local_decl() checked
type modifiers for STATIC instead of declarator.storage_class. For
struct types STATIC was never in the type system, so static local
structs were allocated on the stack. Fixed by checking
declarator.storage_class. This caused CPython's _PyArg_Parser structs
to live on the stack, corrupting the finalization linked list with
dangling pointers and crashing in _PyArg_Fini().

+++ cc/parse/parser.rs
@@ -3691,7 +3691,7 @@ impl Parser<'_> {
     pub(crate) fn eval_const_expr(&self, expr: &Expr) -> Option<i64> {
         match &expr.kind {
             ExprKind::IntLit(val) => Some(*val),
-            ExprKind::CharLit(c) => Some(*c as i64),
+            ExprKind::CharLit(c) => Some(*c as u8 as i8 as i64),
             // Float literals are constant, return truncated value
             ExprKind::FloatLit(val) => Some(*val as i64),
